Revert "Change `.web` files to default and add `.native` (#2835)"
This reverts commit e3e0b79a3d3a622ec65a9c9509e8b5b12b01ff6e.
