+++ src/RNRenderer.ts
@@ -1,3 +1,3 @@
-export const RNRenderer = {
-  findHostInstance_DEPRECATED: (_ref: any) => null,
-};
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
+export { default as RNRenderer } from 'react-native/Libraries/Renderer/shims/ReactNative';
