+++ src/RNRenderer.web.ts
@@ -1,0 +1,3 @@
+export const RNRenderer = {
+  findHostInstance_DEPRECATED: (_ref: any) => null,
+};
