--- src/components/GestureHandlerButton.native.tsx
@@ -1,5 +1,0 @@
-import { HostComponent } from 'react-native';
-import { RawButtonProps } from './GestureButtons';
-import RNGestureHandlerButtonNativeComponent from '../specs/RNGestureHandlerButtonNativeComponent';
-
-export default RNGestureHandlerButtonNativeComponent as HostComponent<RawButtonProps>;
