+++ src/components/GestureHandlerButton.tsx
@@ -1,6 +1,5 @@
-import * as React from 'react';
-import { View } from 'react-native';
+import { HostComponent } from 'react-native';
+import { RawButtonProps } from './GestureButtons';
+import RNGestureHandlerButtonNativeComponent from '../specs/RNGestureHandlerButtonNativeComponent';
 
-export default React.forwardRef<View>((props, ref) => (
-  <View ref={ref} accessibilityRole="button" {...props} />
-));
+export default RNGestureHandlerButtonNativeComponent as HostComponent<RawButtonProps>;
