+++ src/components/GestureHandlerButton.web.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import { View } from 'react-native';
+
+export default React.forwardRef<View>((props, ref) => (
+  <View ref={ref} accessibilityRole="button" {...props} />
+));
