+++ src/getReactNativeVersion.web.ts
@@ -1,0 +1,3 @@
+export function getReactNativeVersion() {
+  throw new Error('getReactNativeVersion is not supported on web');
+}
