--- src/handlers/PressabilityDebugView.native.tsx
@@ -1,2 +1,0 @@
-// @ts-ignore it's not exported so we need to import it from path
-export { PressabilityDebugView } from 'react-native/Libraries/Pressability/PressabilityDebug';
