+++ src/handlers/PressabilityDebugView.tsx
@@ -1,4 +1,2 @@
-// PressabilityDebugView is not implemented in react-native-web
-export function PressabilityDebugView() {
-  return null;
-}
+// @ts-ignore it's not exported so we need to import it from path
+export { PressabilityDebugView } from 'react-native/Libraries/Pressability/PressabilityDebug';
