+++ src/handlers/PressabilityDebugView.web.tsx
@@ -1,0 +1,4 @@
+// PressabilityDebugView is not implemented in react-native-web
+export function PressabilityDebugView() {
+  return null;
+}
