--- src/handlers/customDirectEventTypes.native.ts
@@ -1,2 +1,0 @@
-// @ts-ignore - its taken straight from RN
-export { customDirectEventTypes } from 'react-native/Libraries/Renderer/shims/ReactNativeViewConfigRegistry';
